make DDC brightness changer more resilient
one thing i noticed is that if i try to further increase brightness while at max
the DDC monitor API fails often. I am suspecting that when we try to set brightness
to current brightness, it somehow causes a failure.

Update: I'm still seeing failures with DDC not at the edges, but somewhere in between